requirements/base.txt
Django==4.1.5
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-yasg==1.21.4
python-dotenv==0.21.0
